Avoid low-impact bug in Grid cropping
There seems to be a bug in the Grid's cropping functionality, that in
some circumstances entering an x-coordinate into the cropping form will
result in x-1 being applied. We've been unable to replicate outside of
the cypress environment (and even then, only on the deployed test
runners), and we believe that primary usage of the cropping tool is to
use the UI handles, not the coordinate entry boxes. Picking a different
value for the test's x-coordinate seems to avoid this bug, allowing the
tests to run as usual again.